+++ R/filtering.R
@@ -12,8 +12,11 @@
 #'
 #' If low_freq > high_freq, bandstop filtering is performed.
 #'
-#' Note that the signal is first zero-meaned using either channel means or
-#' by-channel epoch means.
+#' Note that it is recommended to first zero-mean the signal using either
+#' channel means or by-channel epoch means.
+#'
+#' The function allows parallelization using the `future` package, e.g. using
+#' `plan(multiprocess)`
 #'
 #' @section FIR versus IIR filtering: Finite Impulse Response (FIR) filtering is
 #'   performed using an overlap-add FFT method. Note that this only performs a
@@ -46,6 +49,7 @@ eeg_filter <- function(.data, ...) {
 #'   Response). Defaults to "fir".
 #' @param window Windowing function to use (FIR filtering only). Defaults to
 #'   "hamming"; currently only "hamming" available.
+#' @param demean Remove DC component (i.e. channel/epoch mean) before filtering. Defaults to TRUE.
 #' @rdname eeg_filter
 #' @export
 eeg_filter.eeg_data <- function(.data,
@@ -55,6 +59,7 @@ eeg_filter.eeg_data <- function(.data,
                                 trans_bw = "auto",
                                 method = "fir",
                                 window = "hamming",
+                                demean = TRUE,
                                 ...) {
 
   filt_pars <- parse_filt_freqs(low_freq,
@@ -98,7 +103,10 @@ eeg_filter.eeg_data <- function(.data,
                             filter_order,
                             window)
 
-  .data <- rm_baseline(.data) # remove DC component
+  if (demean) {
+    .data <- rm_baseline(.data) # remove DC component
+  }
+
   if (identical(method, "iir")) {
     .data <- run_iir_n(.data,
                        filt_coef)
@@ -121,6 +129,7 @@ eeg_filter.eeg_epochs <- function(.data,
                                   trans_bw = "auto",
                                   method = "fir",
                                   window = "hamming",
+                                  demean = TRUE,
                                   ...) {
 
   filt_pars <- parse_filt_freqs(low_freq,
@@ -164,7 +173,9 @@ eeg_filter.eeg_epochs <- function(.data,
                             filter_order,
                             window)
 
-  .data <- rm_baseline(.data) # remove DC component
+  if (demean) {
+    .data <- rm_baseline(.data) # remove DC component
+  }
   if (identical(method, "iir")) {
     .data <- run_iir_n(.data,
                        filt_coef)
@@ -405,7 +416,7 @@ select_window <- function(type,
                           m,
                           a = NULL) {
 
-    m <- m + 1
+  m <- m + 1
   w <- switch(type,
               "bartlett" = signal::bartlett(m),
               "hann" = signal::hanning(m),