+++ src/RcppExports.cpp
@@ -30,6 +30,21 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+
+// do_iter
+arma::mat do_iter(arma::cube amps, int k, int N);
+RcppExport SEXP _eegUtils_do_iter(SEXP ampsSEXP, SEXP kSEXP, SEXP NSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::cube >::type amps(ampsSEXP);
+    Rcpp::traits::input_parameter< int >::type k(kSEXP);
+    Rcpp::traits::input_parameter< int >::type N(NSEXP);
+    rcpp_result_gen = Rcpp::wrap(do_iter(amps, k, N));
+    return rcpp_result_gen;
+END_RCPP
+}
+
 // get_list
 Rcpp::List get_list(const Rcpp::List& x, const arma::cx_mat& mors, const double n_kern, const std::string output);
 RcppExport SEXP _eegUtils_get_list(SEXP xSEXP, SEXP morsSEXP, SEXP n_kernSEXP, SEXP outputSEXP) {
@@ -76,6 +91,7 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_eegUtils_baseline_cont", (DL_FUNC) &_eegUtils_baseline_cont, 2},
     {"_eegUtils_baseline_epo", (DL_FUNC) &_eegUtils_baseline_epo, 2},
+    {"_eegUtils_do_iter", (DL_FUNC) &_eegUtils_do_iter, 3},
     {"_eegUtils_get_list", (DL_FUNC) &_eegUtils_get_list, 4},
     {"_eegUtils_get_listal", (DL_FUNC) &_eegUtils_get_listal, 4},
     {"_eegUtils_get_list2", (DL_FUNC) &_eegUtils_get_list2, 4},