+++ src/tfr_test.cpp
@@ -9,13 +9,13 @@ Rcpp::List get_list2(const Rcpp::List& x,
                      const std::string output) {
 
   Rcpp::List jj(x.length());
+  Rcpp::Function rfft("mvfft");
   int npoints = mors.n_rows;
   double hmx = n_kern/2;
   int hmz = floor(hmx);
   for (int i = 0; i < x.length(); i++) {
     Rcpp::DataFrame df = Rcpp::as<Rcpp::DataFrame>(x[i]);
     int no_rows = df.nrows();
-    int end_point = no_rows - 1;
     NumericMatrix y(no_rows, df.size());
 
     for (int j = 0; j < df.size() ; j++) {
@@ -23,17 +23,17 @@ Rcpp::List get_list2(const Rcpp::List& x,
     }
 
     arma::mat yk(y.begin(), no_rows, df.size(), false);
-    arma::cx_mat yx(arma::fft(yk, npoints));
+    yk.resize(npoints, df.size());
+    ComplexMatrix yf(rfft(yk));
 
-    int time_end = yx.n_rows - hmz;
+    arma::cx_mat yx(reinterpret_cast<std::complex<double>*>(yf.begin()), npoints, df.size());
+
     int timez = hmz + no_rows - 1;
-    int final_points = time_end - hmz;
 
     arma::cx_cube yz(no_rows, yx.n_cols,
-                  mors.n_cols,
-                  arma::fill::zeros);
+                  mors.n_cols);
 
-    for (arma::uword k = 0; k < yx.n_cols ; k++) {
+    for (arma::uword k = 0; k < yx.n_cols; k++) {
       arma::cx_mat cg(arma::ifft(mors.each_col() % yx.col(k)));
       yz.col(k) = cg.rows(hmz, timez);
     }