+++ NEWS.md
@@ -1,14 +1,19 @@
-# eegUtils 0.4.0.9000
+# eegUtils 0.5.0
 
 ### Function changes
 - Default settings for Infomax ICA changed to be similar to EEGLAB/Fieldtrip.
 - Faster reading of bdf implemented. Old behaviour can be retained using `fast_bdf = FALSE` parameter to `import_raw()`
 - `eeg_combine` now supports combining lists.
 - `eeg_reference` now supports `eeg_epochs` and `eeg_evoked` objects.
-
 ### Internal changes / bug fixes
 - `plot_butterfly` should now be faster again.
 - Much faster reader for BDF implemented.
+- `eeg_filter` added `demean` parameter so that removing channel/epoch means during filtering is now optional. Defaults to TRUE.
+- Added artefact detection options for ICA objects
+    - `ar_acf()` checks for low autocorrelation
+    - `ar_chanfoc()` checks for excessive channel focality (e.g. components that load mostly on one channel)
+    - `ar_trialfoc()` checks for trial focality (components that load mostly on a few trials)
+    - `ar_eogcor()` checks for correlation with EOG channels
 - `topoplot` plotting radius logic altered 
 - `compute_csd` now uses `eeg_reference` rather than `reref_eeg`
 - Unmixing matrix for SSD decompositions fixed.
@@ -20,6 +25,8 @@
 - Updated R requirement to >= 3.2.0
 - Updated rlang requirement to >= 0.4.0
 - `compute_psd` fix for single segment data
+- updated use of `nest` and `unnest` in keeping with `tidyr 1.0.0`
+- `as.data.frame.eeg_tfr()` now fixed to output correctly
 
 # eegUtils 0.4.0